♻️ refactor(Home): Home 화면에 들어가는 항목들 컴포넌트화
모든 페이지에서 쓰이는 최상단 div도 컴포넌트화(Main.jsx)
Issue #6
