+++ client/src/components/Main.jsx
@@ -1,0 +1,5 @@
+import tw from 'tailwind-styled-components';
+
+const Main = tw.div`mx-auto flex h-screen w-screen max-w-[480px] flex-col overflow-hidden bg-paper bg-cover bg-no-repeat`;
+
+export { Main };
