+++ client/src/components/homepage/MemberInfo.jsx
@@ -1,0 +1,14 @@
+const MemberInfo = () => {
+  return (
+    <div className='flex flex-col items-center'>
+      <p id='data' className='text-3xl'>
+        06 / 06
+      </p>
+      <p id='name' className='text-[2.5rem]'>
+        Lee Jae Woong
+      </p>
+    </div>
+  );
+};
+
+export { MemberInfo };
