+++ client/src/components/homepage/index.js
@@ -1,0 +1,4 @@
+export * from './Dot';
+export * from './MemberInfo';
+export * from './Header';
+export * from './EnterButton';
